refactor: move api routes to a specific folder
