+++ .eslintrc.json
@@ -1,5 +1,9 @@
 {
   "extends": "standard-with-typescript",
+  "env": {
+    "browser": true,
+    "node": true
+  },
   "parserOptions": {
     "project": "./tsconfig.json"
   },
