+++ src/main/config/routes.ts
@@ -2,10 +2,10 @@ import { Express, Router } from 'express'
 import fg from 'fast-glob'
 
 export default (app: Express): void => {
-  const router = Router()
-  app.use('/api', router)
+  const apiRouter = Router()
+  app.use('/api', apiRouter)
 
-  fg.sync('**/src/main/routes/**/**routes.ts').map(async (file) => {
-    (await import(`../../../${file}`)).default(router)
+  fg.sync('**/src/main/routes/api/**/**routes.ts').map(async (file) => {
+    (await import(`../../../${file}`)).default(apiRouter)
   })
 }
